mmc: renesas_sdhi: Disabled auto correction in HS400
In R-Car Gen3, auto correction cannot use by HW failure in HS400.

The value sampled by the selected TAP is compared with the value
sampled by previous and after TAP position.
However, comparison HW does not work correctly in HS400.

When comparison result is not match, TAP position is automatically
changed. As a result, TAP moves to the wrong position.

Signed-off-by: Takeshi Saito <takeshi.saito.xv@renesas.com>