mtd: Consolidate Renesas RPC drivers
This consolidates RPC HyperFlash and SPI NOR drivers.
Flash device detection, and DMA read operation is done
by the renesas-rpc driver. The operation mode is selected
based on the child flash device node. If it's "jedec,spi-nor"
compatible, the "renesas-rpc-qspi" platform device is created.
If it's "cfi-flash" compatible, the "renesas-rpc-hyperflash"
device is created instead.

The RPC device compatibility string is adjusted to match U-Boot.
The following configuration options are available:

  CONFIG_MTD_RENESAS_RPC: enables generic RPC support;
  CONFIG_MTD_RENESAS_RPC_QSPI: enables SPI NOR support;
  CONFIG_MTD_RENESAS_RPC_HYPERFLASH: enables HyperFlash support.

Signed-off-by: Valentine Barshak <valentine.barshak@cogentembedded.com>