media: rcar-vin: Add UDS function support
Add UDS (Up Down Scale) control for R-Car Gen3.
UDS can be vertical and horizontal scaling.

There is a possibility that scaling may not be executed
when VUP bit in VnMC register is active and UDS relation registers
can be not set. Therefore, Add processing to set VUP bit to 0 at
the capture stop.

In addition, before stopping the capture, if UDS (Up Down Scale) is set
to disable and waiting for the capture completion, capture data invades
to the area beyond the area allocated by dma_alloc function.
Since it was obtained by the view of the H/W team that it is better to
disable UDS after stopping the capture.

Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>