drm: bridge: hw-hdmi: Control clock of bridge during S2RAM
This patch helps to control the clocks of
bridge/hw-hdmi during system suspend/resume.

Currently, "hdmi" clock is enabled when the DU driver
is enabled but it is not disabled at system suspend.

So, when the system resumes the software state (enabled)
and hardware (disabled) of clock are not matched.
And it makes the system resume fails.

This patches is necessary to synchronize software
and hardware state of clock when backup/restore code
of clock (CPG module) is removed.

In addtion, the procedure for enabling hot plug is added.

Reported-by: Dien Pham <dien.pham.ry@renesas.com>
Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>
Signed-off-by: LUU HOAI <hoai.luu.ub@renesas.com>