clk: renesas: r8a77{95, 96, 965, 990}: Remove INTC-AP and INTC-EX clocks
According to the R-Car Gen3 Hardware Manual Rev 1.00, the module clock
status, module clock control, and reset control bits of the INTC-AP and
INTC-EX modules on R-Car H3-W, R-Car M3, R-Car M3-N and R-Car E3 are removed.

Continue to change the clock properties of the gic and irqc nodes in the
device tree.

Signed-off-by: Takeshi Kihara <takeshi.kihara.df@renesas.com>
Signed-off-by: LUU HOAI <hoai.luu.ub@renesas.com>