spi: sh-msiof: Add sleep before master transfer for test
This patch is for debug of transfer between master and slave.
Since the slave needs to complete a preparation in data transfer
before the master working, the sleep wait is put before
the data transfer of the master.

Signed-off-by: Ryo Kataoka <ryo.kataoka.wt@renesas.com>
(cherry picked from commit 0350e7269fb88881a33dc7327852c46a71e3ad13)