tee: tee_shm: Fix the release function to the shared memory
When the kernel module releases the shared memory by the
tee_shm_free function, this function performs setting for
releasing at the end of the task but does not actually release the
memory. Therefore, when securing a lot of memory with one task, an
error occurs, so the shared memory is released each time
tee_shm_free is called.

Fixed checkpatch issues.

Signed-off-by: Takeshi Kihara <takeshi.kihara.df@renesas.com>