i2c: i2c-rcar: Add delay W/A for R8A779A0
In R8A779A0 device, the interrupt handler does not work well
due to unknown reason. After testing with various case with
this interrupt handler, we found that the interrupt can work well
if we added a delay inside of it. To solve this problem, we add
a temporary delay to let interrupt handler can work.

Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>