i2c: rcar: Fix handler processing when unintended interrupt
In R8A779A0, even though MDT interrupts are disabled,
only MDT status may be observed in the interrupt handler.

In that case, if the interrupt handler is called only by
MDT status and the process (Clear START or STOP immediately)
in the handler is executed, the I2C transfer will fail.

This patch modifies this process so that it will not be
executed when an unintended interrupt occurs.

Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>