remoteproc: rcar: If the CR7 FW is already running, leave it alone
Most use cases for the Cortex-R7 require it to be running long before
linux is started. Therefore, if the CR7 is already running, do not
reset it and do not download new firmware.

The driver still requires the elf firmware in the rootfs, as this is used
to get the virtio resource table information.

Prior to this change, the driver relied on the "memory-region" being part
of CMA, and that the CR7 virtio resource table had a carveout region for
the whole of DDR used by the CR7.

Now, the "memory-region" is not part of CMA. The virtio resource table
specifies two 512 byte vrings (4kB aligned) for rpmsg which are allocated
out of the existing CMA region, i.e. in the address range 0x58000000 to
0x5c000000, not in the address range 0x70000000 to 0x80000000.
This means the CR7 firmware must no longer specify the carveout for it's
code and data.

Signed-off-by: Phil Edworthy <phil.edworthy@renesas.com>