memory: renesas-rpc-if: Correct data transfer in Manual mode
This patch fixes the output warning logs and data loss when
performing mount/umount then remount the device with jffs2 format.

This is the sample warning logs when performing mount/umount then
remount the device with jffs2 format:
jffs2: jffs2_scan_inode_node(): CRC failed on node at 0x031c51d4:
Read 0x00034e00, calculated 0xadb272a7

The reason for this issue is that the writing data seems to get
messed up.
Data is only completed when the number of bytes is divisible by 4.
If you only have 3 bytes of data left to write, 1 garbage byte
is inserted after the end of the write stream.
If you only have 2 bytes of data left to write, 2 bytes of '00'
are added into the write stream.
If you only have 1 byte of data left to write, 2 bytes of '00'
are added into the write stream. 1 garbage byte is inserted after
the end of the write stream.

To solve this problem, data must be written continuously in serial
and the write stream ends when data is out.

Stop using the 'memcpy' function to copy block of memory when
writing data in Manual mode to avoid alignment problem.

Macro 'get_unaligned' should be used for accessing values larger
in size than single bytes at locations that are expected to be
improperly aligned.
E.g. Retrieving a u32 value from a location not u32-aligned.

Signed-off-by: Duc Nguyen <duc.nguyen.ub@renesas.com>