arm64: dts: salvator-common: Add no-sdio and no-sd properties to SDHI2
Add the 'no-sdio' and 'no-sd' properties to the SDHI2 node used as the
eMMC port on Salvator board. It will limits the send sdio and sd cmd
during mmc initialization.

In addition, the 'no-sdio' property uses multiple SG entries when SDHI
internal DMAC with IOMMU is enabled. This can improve performance.

Signed-off-by: Takeshi Saito <takeshi.saito.xv@renesas.com>
Signed-off-by: Takeshi Kihara <takeshi.kihara.df@renesas.com>
Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>