arm64: dts: r8a77990-ebisu{,-es10}: Add no-sdio and no-sd properties to SDHI3
Add the 'no-sdio' and 'no-sd' properties to the SDHI3 node used as the
eMMC port on Ebisu and Ebisu-4D boards. It will limits the send sdio and
sd cmd during mmc initialization.

In addition, the 'no-sdio' property uses multiple SG entries when SDHI
internal DMAC with IOMMU is enabled. This can improve performance.

Signed-off-by: Takeshi Saito <takeshi.saito.xv@renesas.com>
Signed-off-by: Takeshi Kihara <takeshi.kihara.df@renesas.com>
Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>