soc: renesas: rcar-sysc: Include IMP power domains as always-on power domains
As in commit (soc: renesas: rcar-sysc: Improve power
domains always on implementation (1*)), it shows that
Hardware needs to be controlled with the state transitions
do not conflict for each power domains, especially with
some power domains, which are transitioned by multiple
triggers.

This patch will implement this point for all IMP power domains.

Signed-off-by: Tho Vu <tho.vu.wh@renesas.com>
[updated for supporting R-Car V3H 1.x]
Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>