mmc: tmio: always flag retune when resetting and a card is present
After reset, we manually flagged retune in runtime resume, but missed it
in the workqueue. To fix that and avoid the problem in the future, let's
flag retune in the reset handler directly whenever a card is present.

Reported-by: Takeshi Saito <takeshi.saito.xv@renesas.com>
Signed-off-by: Wolfram Sang <wsa+renesas@sang-engineering.com>
Patchwork: https://patchwork.kernel.org/project/linux-mmc/patch/20210316085717.7276-3-wsa+renesas@sang-engineering.com/
Signed-off-by: Takeshi Saito <takeshi.saito.xv@renesas.com>