arm64: dts: renesas: condor: switch eMMC bus to 1V8
The eMMC card has two supplies, VCC and VCCQ. The VCC supplies the NAND
array and the VCCQ supplies the bus. On Condor, the VCC is connected to
3.3V rail, while the VCCQ is connected to 1.8V rail. Adjust the pinmux
to match the bus, which is always operating in 1.8V mode.

While at it, deduplicate the pinmux entries, which are now the same for
both default and UHS modes. We still need the two pinctrl entries to
match the bindings though.

Thanks to Marek Vasut for this commit description from 5f65328df3f5.

Signed-off-by: Wolfram Sang <wsa+renesas@sang-engineering.com>
Reviewed-by: Geert Uytterhoeven <geert+renesas@glider.be>

Link:https://patchwork.kernel.org/project/linux-renesas-soc/patch/20210419143858.39401-1-wsa+renesas@sang-engineering.com/

Signed-off-by: LUU HOAI <hoai.luu.ub@renesas.com>