media: rcar-vin: Fix reset and release procedure
Currently, the UIO driver and the native kernel are
competing for reset processing
(Warning is occurred in the Kernel I/F if the reset resource is the same).
The user chooses the UIO or NativeKernel driver.
so, in order to be able to use reset in either case, the native kernel
side temporarily implements reset processing using ioremap.

In this case, it was only ISP, so only ISP is corresponded.

Signed-off-by: Koji Matsuoka <koji.matsuoka.xm@renesas.com>
Signed-off-by: LUU HOAI <hoai.luu.ub@renesas.com>