thermal: rcar_gen3_thermal: Do not use interrupts for normal operation
Remove the usage of interrupts for the normal temperature operation and
depend on the polling performed by the thermal core. This is done to
prepare to use the interrupts as they are intended to trigger once
specific trip points are passed and not to react to temperature changes
in the normal operational range.

Signed-off-by: Niklas Soderlund <niklas.soderlund+renesas@ragnatech.se>
Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>
(cherry picked from commit 1b57b9597c779f0a20c5837abec31f600240496b)
Signed-off-by: Linh Phung <linh.phung.jy@renesas.com>