media: rcar-vin: Fix the init process of VIN driver
The current init process is not correct because the last device
is opened while it's data struct member is not complete (ctrl_handler = NULL).
This leads kernel hang up when insmod vin module. This patch fix the
init process to make sure that all device's data is initialized
before open them.

Signed-off-by: LUU HOAI <hoai.luu.ub@renesas.com>