arm64: dts: renesas: r8a779f0: Disable legacy DMAE driver
This patch to set Legacy's DMA Engine to disable because UIO and
interrupt conflicts. In the case, if you want to enable it,
set rtdmac_control_xx (xx: 00-03) and irq_rtdmac_control_yy_zz
(with yy: 00-003; zz: 01-15) on the UIO side to status = "disabled"
as default and remove status = "disabled" on native DMAE side.

Signed-off-by: Tam Nguyen <tam.nguyen.xa@renesas.com>