memory: renesas-rpc-if: Clear HS bit during hardware initialization
According to the datasheet, HS bit should be specified to 1 when using
DMA transfer. As DMA transfer is not supported, it should be cleared to 0

Previously, the driver relies on HS bit setting in prior firmware but
this is not always the case.

Fix this by ensuring the bit is cleared during hardware initialization.

Signed-off-by: Cong Dang <cong.dang.xn@renesas.com>
Signed-off-by: Hai Pham <hai.pham.ud@renesas.com>