arm64: defconfig: Enable loadable EMS/IPA module.
Currently, EMS module is implemented as built-in module.
At first thinking, EMS is always necessary to keep
the system running safely (avoid so high temperature.)

But in some cases, customer has high demand to reduce
the kernel binary image size and they request to
build EMS as loadable module.

This patch is implemented for above request.

Signed-off-by: Duy Nguyen <duy.nguyen.rh@renesas.com>