ravb: Issue gpio reset to AVB1/AVB2 PHYs
R-Car V4H(R8A779G0) AVB1/AVB2 PHY is left at an invalid state
during system power-up or reset thus causing unreliability
issue with PHY which manifests as PHY not being detected.
To fix this, these PHYs need to be reset via a GPIO connected
to the PHY's reset pin.

In this case, the reset de-assertion can not be done in the PHY
driver as the PHY will not probe till its reset is de-asserted.
So, the reset should be done before MDIO bus register to make
registration successfully.

Signed-off-by: Tho Vu <tho.vu.wh@renesas.com>
Signed-off-by: Nguyen Tran <nguyen.tran.pz@bp.renesas.com>