arm64: dts: renesas: r8a77980: Update ICCOM UIO devices
Currently, ICCOM UIO devices are using device nodes with
ICCOM native, so the device needs to be separate. When
handling this, the ICCOM native devices must be disabled
to avoid conflict with UIO devices.

The patch also removes ICCOM UIO devices that were redundant
and then they were updated the same as V4H devices.

Signed-off-by: Tam Nguyen <tam.nguyen.xa@renesas.com>