PCI: dwc: rcar-gen4-host: Fix kernel panic when PCIe link is downgraded
Sometimes, because the environment is not good (quality of board, quality of
the oculink cable, connection between the board and oculink cable) will affect
the PCIe link leading to it not reaching to their specification 16GT/s x2lanes
and cause panic.

Using mdelay() instead of msleep() to solve this issue.

Signed-off-by: Tho Vu <tho.vu.wh@renesas.com>
Signed-off-by: Ryo Araki <ryo.araki.uh@renesas.com>
Signed-off-by: Tin Tran <tin.tran.xk@renesas.com>