i2c: rcar: add fix for condition checking to run DMA continuous mode
For I2C DMA continuous mode, still has some restriction as below:
[1] Wrong checking condition for support DMA continuous mode: all transfers
longer than 16 bytes will be automatically run using this mode,
even though it is not actually required.

[2] Wrong define transfer size: assigning the default transfer size
to 4 bytes if data cannot be retrieved from the device tree is wrong.
If true, it would be advisable not to run this mode
if the dma transfer size is not defined in the device tree.

[3] Wrong data transferred when number of descriptor is large:if define
1 transfer size in device tree is small, but the length of transmitted
message is large. The number of descriptors created will be large
and the data transmitted will be incorrect.

This patch apply to fix the restriction [1].

Signed-off-by: Minh Le <minh.le.aj@renesas.com>