mmc: renesas_sdhi: Add reset flow for tuning error
As updated in the H/W Errata for R-Car Series 3rd Generation, the
following steps will be executed before IPMMU unmapping in the event
of an error, in order to prevent IPMMU errors. Those steps are:
- Reset SDHI controller
- Abort DMA
There are also additional steps to be executed but not required before
unmapping:
- Set bus width
- Set clock

Because IPMMU unmapping is implemented in a tasklet, usleep_range()
is replaced by mdelay() when resetting SDHI controller. These changes
are applied only if tuning is error in order not to impact the current
implementation as mdelay() will cause CPU wasting time.

Moreover, setting clock flow eventually might call the mutex_lock()
function which isn't allowed in a tasklet. For example:

host->set_clock() = renesas_sdhi_set_clock()
----> renesas_sdhi_clk_update()
--------> clk_round_rate()
------------> clk_prepare_lock()
----------------> mutex_lock().

So this step will be called in the tmio_mmc_finish_request() function.

Signed-off-by: Vinh Nguyen <vinh.nguyen.xz@renesas.com>