net: ethernet: renesas: ravb: Update Ethernet-AVB driver to align with HW UM
This patch is purely an alignment update to match the HW UM and
does not affect Ethernet-AVB functionality.

According to the current HW UM:

The ECSIPR_ICDIP bit is only valid at 10Mbps/100Mbps. In the actual
RTL design, this bit is disabled in 1000Mbps mode, meaning that even if
1 is written, it functions as 0. The HW UM has been updated to reflect
this design accurately.

The EIS_QFS bit has been corrected from R/W to Read-Only due to a
typographical error. As a result, writing a value to this bit has
no effect.

Although these changes do not impact software operation, updating
the driver is necessary to maintain consistency with the current HW UM.

Signed-off-by: Thanh Quan <thanh.quan.xn@renesas.com>
Signed-off-by: Nguyen Tran <nguyen.tran.pz@bp.renesas.com>