soc: renesas: r8a779g0-sysc: Remove A3VIP1 power domain
From V4H HWUM Errata List 250214, A3VIP1 and A1CNN0 power domains
are removed.
This commit removes the A3VIP1 only because:
- A3VIP1 isn't used by any device until now.
- A1CNN0 is controlled by power domain framework for r8a779g0 ES1.*
while for the others, it is set as PD_FORCE_ALWAYS_ON which makes it
is not controlled by power domain framework. Hence it is not removed.

Signed-off-by: Vinh Nguyen <vinh.nguyen.xz@renesas.com>