gpu: drm: rcar-vcon: Validate mode against dot clock supply
On R-Car X5H v1.0, the PLL clock frequency is relaxed, causing
the dot clock supply frequency to be lower than the documented
value. This can lead to modes being set with unsupported
frequencies.

Add a check to ensure that the dot clock supplied to the panel
does not exceed the hardware limitations.

Signed-off-by: Phong Hoang <phong.hoang.wz@renesas.com>