media: rcar-vin: rcar-csi2: Add 1/2/3/4 lane support for MIPI CSI-2 C-PHY
The current driver only supports 4-lane data mode, while the HW is capable of 1/2/3/4 data lanes.
Based on the customer’s request, we have added support for 1/2/3/4 data lane configurations.

This requires the customer to define the corresponding lane count in the Device Tree (DT) in order to use it.
Only 2/4 data lane configurations is authenticated.

Signed-off-by: Linh Phung <linh.phung.jy@renesas.com>